break out compare.ts from metamodel.ts (using an agent)
